feat(data): make monster level ushort

+++ Data/NosSmooth.Data.Abstractions/Infos/IMonsterInfo.cs
@@ -21,5 +21,5 @@ public interface IMonsterInfo : IVNumInfo
     /// <summary>
     /// Gets the default level of the monster.
     /// </summary>
-    int Level { get; }
+    ushort Level { get; }
 }
